Data.Array.Repa.Operators.Traversal

Documentation

traverse

Arguments

:: (Source r a, Shape sh)

=> Array r sh a
Source array.

-> (sh -> sh')
Function to produce the extent of the result.

-> ((sh -> a) -> sh' -> b)
Function to produce elements of the result. It is passed a lookup function to get elements of the source.

-> Array D sh' b

Unstructured traversal.

unsafeTraverse

Arguments

:: (Source r a, Shape sh)

=> Array r sh a
Source array.

-> (sh -> sh')
Function to produce the extent of the result.

-> ((sh -> a) -> sh' -> b)
Function to produce elements of the result. It is passed a lookup function to get elements of the source.

-> Array D sh' b

Unstructured traversal.

traverse2

Arguments

:: (Source r1 a, Source r2 b, Shape sh, Shape sh')

=> Array r1 sh a
First source array.

-> Array r2 sh' b
Second source array.

-> (sh -> sh' -> sh'')
Function to produce the extent of the result.

-> ((sh -> a) -> (sh' -> b) -> sh'' -> c)
Function to produce elements of the result. It is passed lookup functions to get elements of the source arrays.

-> Array D sh'' c

Unstructured traversal over two arrays at once.

unsafeTraverse2

Arguments

:: (Source r1 a, Source r2 b, Shape sh, Shape sh')

=> Array r1 sh a
First source array.

-> Array r2 sh' b
Second source array.

-> (sh -> sh' -> sh'')
Function to produce the extent of the result.

-> ((sh -> a) -> (sh' -> b) -> sh'' -> c)
Function to produce elements of the result. It is passed lookup functions to get elements of the source arrays.

-> Array D sh'' c

Unstructured traversal over two arrays at once.

traverse3 :: forall r1 r2 r3 sh1 sh2 sh3 sh4 a b c d. (Source r1 a, Source r2 b, Source r3 c, Shape sh1, Shape sh2, Shape sh3) => Array r1 sh1 a -> Array r2 sh2 b -> Array r3 sh3 c -> (sh1 -> sh2 -> sh3 -> sh4) -> ((sh1 -> a) -> (sh2 -> b) -> (sh3 -> c) -> sh4 -> d) -> Array D sh4 d

Unstructured traversal over three arrays at once.

unsafeTraverse3 :: forall r1 r2 r3 sh1 sh2 sh3 sh4 a b c d. (Source r1 a, Source r2 b, Source r3 c, Shape sh1, Shape sh2, Shape sh3) => Array r1 sh1 a -> Array r2 sh2 b -> Array r3 sh3 c -> (sh1 -> sh2 -> sh3 -> sh4) -> ((sh1 -> a) -> (sh2 -> b) -> (sh3 -> c) -> sh4 -> d) -> Array D sh4 d

Unstructured traversal over three arrays at once.

traverse4 :: forall r1 r2 r3 r4 sh1 sh2 sh3 sh4 sh5 a b c d e. (Source r1 a, Source r2 b, Source r3 c, Source r4 d, Shape sh1, Shape sh2, Shape sh3, Shape sh4) => Array r1 sh1 a -> Array r2 sh2 b -> Array r3 sh3 c -> Array r4 sh4 d -> (sh1 -> sh2 -> sh3 -> sh4 -> sh5) -> ((sh1 -> a) -> (sh2 -> b) -> (sh3 -> c) -> (sh4 -> d) -> sh5 -> e) -> Array D sh5 e

Unstructured traversal over four arrays at once.

unsafeTraverse4 :: forall r1 r2 r3 r4 sh1 sh2 sh3 sh4 sh5 a b c d e. (Source r1 a, Source r2 b, Source r3 c, Source r4 d, Shape sh1, Shape sh2, Shape sh3, Shape sh4) => Array r1 sh1 a -> Array r2 sh2 b -> Array r3 sh3 c -> Array r4 sh4 d -> (sh1 -> sh2 -> sh3 -> sh4 -> sh5) -> ((sh1 -> a) -> (sh2 -> b) -> (sh3 -> c) -> (sh4 -> d) -> sh5 -> e) -> Array D sh5 e

Unstructured traversal over four arrays at once.