Data.Array.Repa.Repr.Cursored

Documentation

data C

Cursored Arrays. These are produced by Repa's stencil functions, and help the fusion framework to share index compuations between array elements.

The basic idea is described in ``Efficient Parallel Stencil Convolution'', Ben Lippmeier and Gabriele Keller, Haskell 2011 -- though the underlying array representation has changed since this paper was published.

makeCursored

Arguments

:: sh

-> (sh -> cursor)
Create a cursor for an index.

-> (sh -> cursor -> cursor)
Shift a cursor by an offset.

-> (cursor -> e)
Compute the element at the cursor.

-> Array C sh e

Define a new cursored array.