Data.Array.Repa.Repr.HintSmall

Documentation

data S r1

Hints that evaluating this array is only a small amount of work. It will be evaluated sequentially in the main thread, instead of in parallel on the gang. This avoids the associated scheduling overhead.

hintSmall :: Array r1 sh e -> Array (S r1) sh e

Wrap an array with a smallness hint.