Data.Array.Repa.Repr.Partitioned

Documentation

data P r1 r2

Partitioned arrays. The last partition takes priority

These are produced by Repa's support functions and allow arrays to be defined using a different element function for each partition.

The basic idea is described in ``Efficient Parallel Stencil Convolution'', Ben Lippmeier and Gabriele Keller, Haskell 2011 -- though the underlying array representation has changed since this paper was published.

data Range sh

Constructors

Range !sh !sh (sh -> Bool)

inRange :: Range sh -> sh -> Bool

Check whether an index is within the given range.