Data.Array.Repa.Arbitrary

Arbitrary Unboxed Arrays

arbitraryUShaped :: (Arbitrary a, Unbox a, Shape sh) => sh -> Gen (Array U sh a)

Generates a random unboxed array of a given shape

forAllUShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Unbox a, Shape sh) => sh -> (Array U sh a -> prop) -> Property

Property tested for unboxed random arrays with a given shape.

forAll2UShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Unbox a, Shape sh) => sh -> ((Array U sh a, Array U sh a) -> prop) -> Property

Property tested for pair of unboxed random arrays with a given shape.

forAll3UShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Unbox a, Shape sh) => sh -> ((Array U sh a, Array U sh a, Array U sh a) -> prop) -> Property

Property tested for triple of unboxed random arrays with a given shape.

forAll4UShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Unbox a, Shape sh) => sh -> ((Array U sh a, Array U sh a, Array U sh a, Array U sh a) -> prop) -> Property

Property tested for quadruple of unboxed random arrays with a given shape.

forAll5UShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Unbox a, Shape sh) => sh -> ((Array U sh a, Array U sh a, Array U sh a, Array U sh a, Array U sh a) -> prop) -> Property

Property tested for 5-tuple of unboxed random arrays with a given shape.

Arbitrary Boxed Arrays

arbitraryVShaped :: (Arbitrary a, Shape sh) => sh -> Gen (Array V sh a)

Generates a random boxed array of a given shape

forAllVShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Shape sh) => sh -> (Array V sh a -> prop) -> Property

Property tested for unboxed random arrays with a given shape.

forAll2VShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Shape sh) => sh -> ((Array V sh a, Array V sh a) -> prop) -> Property

Property tested for pair of unboxed random arrays with a given shape.

forAll3VShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Shape sh) => sh -> ((Array V sh a, Array V sh a, Array V sh a) -> prop) -> Property

Property tested for triple of unboxed random arrays with a given shape.

forAll4VShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Shape sh) => sh -> ((Array V sh a, Array V sh a, Array V sh a, Array V sh a) -> prop) -> Property

Property tested for quadruple of unboxed random arrays with a given shape.

forAll5VShaped :: (Show a, Show sh, Testable prop, Arbitrary a, Shape sh) => sh -> ((Array V sh a, Array V sh a, Array V sh a, Array V sh a, Array V sh a) -> prop) -> Property

Property tested for 5-tuple of unboxed random arrays with a given shape.