Data.Array.Repa.Eval

Description

Low level interface to parallel array filling operators.

Element types

class Elt a where

Element types that can be used with the blockwise filling functions.

This class is mainly used to define the touch method. This is used internally in the imeplementation of Repa to prevent let-binding from being floated inappropriately by the GHC simplifier. Doing a seq sometimes isn't enough, because the GHC simplifier can erase these, and still move around the bindings.

Minimal complete definition

Nothing

Methods

touch :: a -> IO ()

Place a demand on a value at a particular point in an IO computation.

zero :: a

Generic zero value, helpful for debugging.

one :: a

Generic one value, helpful for debugging.

Parallel array filling

class Target r e where

Class of manifest array representations that can be constructed in parallel.

Associated Types

data MVec r e

Mutable version of the representation.

Methods

newMVec :: Int -> IO (MVec r e)

Allocate a new mutable array of the given size.

unsafeWriteMVec :: MVec r e -> Int -> e -> IO ()

Write an element into the mutable array.

unsafeFreezeMVec :: sh -> MVec r e -> IO (Array r sh e)

Freeze the mutable array into an immutable Repa array.

deepSeqMVec :: MVec r e -> a -> a

Ensure the strucure of a mutable array is fully evaluated.

touchMVec :: MVec r e -> IO ()

Ensure the array is still live at this point. Needed when the mutable array is a ForeignPtr with a finalizer.

class (Source r1 e, Shape sh) => Load r1 sh e where

Compute all elements defined by an array and write them to a manifest target representation.

Note that instances require that the source array to have a delayed representation such as D or C. If you want to use a pre-existing manifest array as the source then delay it first.

Methods

loadS :: Target r2 e => Array r1 sh e -> MVec r2 e -> IO ()

Fill an entire array sequentially.

loadP :: Target r2 e => Array r1 sh e -> MVec r2 e -> IO ()

Fill an entire array in parallel.

class (Source r1 e, Shape sh) => LoadRange r1 sh e where

Compute a range of elements defined by an array and write them to a fillable representation.

Methods

loadRangeS :: Target r2 e => Array r1 sh e -> MVec r2 e -> sh -> sh -> IO ()

Fill a range of an array sequentially.

loadRangeP :: Target r2 e => Array r1 sh e -> MVec r2 e -> sh -> sh -> IO ()

Fill a range of an array in parallel.

fromList :: (Shape sh, Target r e) => sh -> [e] -> Array r sh e

O(n). Construct a manifest array from a list.

Converting between representations

computeS :: (Load r1 sh e, Target r2 e) => Array r1 sh e -> Array r2 sh e

Sequential computation of array elements.

computeP :: (Load r1 sh e, Target r2 e, Source r2 e, Monad m) => Array r1 sh e -> m (Array r2 sh e)

Parallel computation of array elements.

The source array must have a delayed representation like D, C or P, and the result a manifest representation like U or F.

If you want to copy data between manifest representations then use copyP instead.

If you want to convert a manifest array back to a delayed representation then use delay instead.

suspendedComputeP :: (Load r1 sh e, Target r2 e) => Array r1 sh e -> Array r2 sh e

Suspended parallel computation of array elements.

This version creates a thunk that will evaluate the array on demand. If you force it when another parallel computation is already running then you will get a runtime warning and evaluation will be sequential. Use deepSeqArray and now to ensure that each array is evaluated before proceeding to the next one.

If unsure then just use the monadic version computeP. This one ensures that each array is fully evaluated before continuing.

copyS :: (Source r1 e, Load D sh e, Target r2 e) => Array r1 sh e -> Array r2 sh e

Sequential copying of arrays.

copyP :: (Source r1 e, Source r2 e, Load D sh e, Target r2 e, Monad m) => Array r1 sh e -> m (Array r2 sh e)

Parallel copying of arrays.

This is a wrapper that delays an array before calling computeP.

You can use it to copy manifest arrays between representations.

suspendedCopyP :: (Source r1 e, Load D sh e, Target r2 e) => Array r1 sh e -> Array r2 sh e

Suspended parallel copy of array elements.

now :: (Shape sh, Source r e, Monad m) => Array r sh e -> m (Array r sh e)

Monadic version of deepSeqArray.

Forces an suspended array computation to be completed at this point in a monadic computation.

 do  let arr2 = suspendedComputeP arr1
     ...
     arr3 <- now $ arr2
     ...

Chunked filling

fillLinearS

Arguments

:: Int
Number of elements.

-> (Int -> a -> IO ())
Update function to write into result buffer.

-> (Int -> a)
Fn to get the value at a given index.

-> IO ()

Fill something sequentially.

The array is filled linearly from start to finish.

fillChunkedP

Arguments

:: Int
Number of elements.

-> (Int -> a -> IO ())
Update function to write into result buffer.

-> (Int -> a)
Fn to get the value at a given index.

-> IO ()

Fill something in parallel.

The array is split into linear chunks, and each thread linearly fills one chunk.

fillChunkedIOP

Arguments

:: Int
Number of elements.

-> (Int -> a -> IO ())
Update fn to write into result buffer.

-> (Int -> IO (Int -> IO a))
Create a fn to get the value at a given index. The first Int is the thread number, so you can do some per-thread initialisation.

-> IO ()

Fill something in parallel, using a separate IO action for each thread.

The array is split into linear chunks, and each thread linearly fills one chunk.

Interleaved filling

fillInterleavedP

Arguments

:: Int
Number of elements.

-> (Int -> a -> IO ())
Update function to write into result buffer.

-> (Int -> a)
Fn to get the value at a given index.

-> IO ()

Fill something in parallel.

The array is split into linear chunks and each thread fills one chunk.

Blockwise filling

fillBlock2P

Arguments

:: Elt a

=> (Int -> a -> IO ())
Update function to write into result buffer.

-> (DIM2 -> a)
Function to evaluate the element at an index.

-> Int#
Width of the whole array.

-> Int#
x0 lower left corner of block to fill

-> Int#
y0

-> Int#
w0 width of block to fill.

-> Int#
h0 height of block to fill.

-> IO ()

Fill a block in a rank-2 array in parallel.

Blockwise filling can be more cache-efficient than linear filling for rank-2 arrays.

Coordinates given are of the filled edges of the block.

We divide the block into columns, and give one column to each thread.

Each column is filled in row major order from top to bottom.

fillBlock2S

Arguments

:: (Int -> a -> IO ())
Update function to write into result buffer.

-> (DIM2 -> a)
Fn to get the value at the given index.

-> Int#
Width of the whole array.

-> Int#
x0 lower left corner of block to fill.

-> Int#
y0

-> Int#
w0 width of block to fill

-> Int#
h0 height of block to fill

-> IO ()

Fill a block in a rank-2 array, sequentially.

Blockwise filling can be more cache-efficient than linear filling for rank-2 arrays.

The block is filled in row major order from top to bottom.

Cursored blockwise filling

fillCursoredBlock2S

Arguments

:: Elt a

=> (Int -> a -> IO ())
Update function to write into result buffer.

-> (DIM2 -> cursor)
Make a cursor to a particular element.

-> (DIM2 -> cursor -> cursor)
Shift the cursor by an offset.

-> (cursor -> a)
Function to evaluate an element at the given index.

-> Int#
Width of the whole array.

-> Int#
x0 lower left corner of block to fill.

-> Int#
y0

-> Int#
w0 width of block to fill

-> Int#
h0 height of block to fill

-> IO ()

Fill a block in a rank-2 array, sequentially.

Blockwise filling can be more cache-efficient than linear filling for rank-2 arrays.

Using cursor functions can help to expose inter-element indexing computations to the GHC and LLVM optimisers.

Coordinates given are of the filled edges of the block.

The block is filled in row major order from top to bottom.

fillCursoredBlock2P

Arguments

:: Elt a

=> (Int -> a -> IO ())
Update function to write into result buffer.

-> (DIM2 -> cursor)
Make a cursor to a particular element.

-> (DIM2 -> cursor -> cursor)
Shift the cursor by an offset.

-> (cursor -> a)
Function to evaluate the element at an index.

-> Int#
Width of the whole array.

-> Int#
x0 lower left corner of block to fill

-> Int#
y0

-> Int#
w0 width of block to fill

-> Int#
h0 height of block to fill

-> IO ()

Fill a block in a rank-2 array in parallel.

Blockwise filling can be more cache-efficient than linear filling for rank-2 arrays.

Using cursor functions can help to expose inter-element indexing computations to the GHC and LLVM optimisers.

Coordinates given are of the filled edges of the block.

We divide the block into columns, and give one column to each thread.

Each column is filled in row major order from top to bottom.

Chunked selection

selectChunkedS

Arguments

:: Shape sh

=> (sh -> a -> IO ())
Update function to write into result.

-> (sh -> Bool)
See if this predicate matches.

-> (sh -> a)
.. and apply fn to the matching index

-> sh
Extent of indices to apply to predicate.

-> IO Int
Number of elements written to destination array.

Select indices matching a predicate.

This primitive can be useful for writing filtering functions.

selectChunkedP

Arguments

:: Unbox a

=> (Int -> Bool)
See if this predicate matches.

-> (Int -> a)

-> Int

-> IO [IOVector a]

Select indices matching a predicate, in parallel.

This primitive can be useful for writing filtering functions.

The array is split into linear chunks, with one chunk being given to each thread.

The number of elements in the result array depends on how many threads you're running the program with.