Data.Array.Repa.Index

Description

Index types.

Index types

data Z

An index of dimension zero

Constructors

Z

data tail :. head infixl 3

Our index type, used for both shapes and indices.

Constructors

!tail :. !head infixl 3

Common dimensions.

type DIM0 = Z

type DIM1 = DIM0 :. Int

type DIM2 = DIM1 :. Int

type DIM3 = DIM2 :. Int

type DIM4 = DIM3 :. Int

type DIM5 = DIM4 :. Int

ix1 :: Int -> DIM1

Helper for index construction.

Use this instead of explicit constructors like (Z :. (x :: Int)). The this is sometimes needed to ensure that x is constrained to be in Int.

ix2 :: Int -> Int -> DIM2

ix3 :: Int -> Int -> Int -> DIM3

ix4 :: Int -> Int -> Int -> Int -> DIM4

ix5 :: Int -> Int -> Int -> Int -> Int -> DIM5