Data.Array.Repa.Repr.Undefined

Documentation

data X

An array with undefined elements.

This is normally used as the last representation in a partitioned array, as the previous partitions are expected to provide full coverage.