Data.Array.Repa.Repr.Vector

Documentation

data V

Arrays represented as boxed vectors.

This representation should only be used when your element type doesn't have an Unbox instsance. If it does, then use the Unboxed U representation will be faster.

computeVectorS :: (Shape sh, Load r1 sh e) => Array r1 sh e -> Array V sh e

Sequential computation of array elements.

This is an alias for compute with a more specific type.

computeVectorP :: (Shape sh, Load r1 sh e, Monad m) => Array r1 sh e -> m (Array V sh e)

Parallel computation of array elements.

fromListVector :: Shape sh => sh -> [a] -> Array V sh a

O(n). Convert a list to a boxed vector array.

This is an alias for fromList with a more specific type.

fromVector :: Shape sh => sh -> Vector e -> Array V sh e

O(1). Wrap a boxed vector as an array.

toVector :: Array V sh e -> Vector e

O(1). Unpack a boxed vector from an array.