Data.Array.Repa.Shape

Description

Class of types that can be used as array shapes and indices.

Documentation

class Eq sh => Shape sh where

Class of types that can be used as array shapes and indices.

Minimal complete definition

rank, zeroDim, unitDim, intersectDim, addDim, size, sizeIsValid, toIndex, fromIndex, inShapeRange, listOfShape, shapeOfList, deepSeq

Methods

rank :: sh -> Int

Get the number of dimensions in a shape.

zeroDim :: sh

The shape of an array of size zero, with a particular dimensionality.

unitDim :: sh

The shape of an array with size one, with a particular dimensionality.

intersectDim :: sh -> sh -> sh

Compute the intersection of two shapes.

addDim :: sh -> sh -> sh

Add the coordinates of two shapes componentwise

size :: sh -> Int

Get the total number of elements in an array with this shape.

sizeIsValid :: sh -> Bool

Check whether this shape is small enough so that its flat indices an be represented as Int. If this returns False then your array is too big. Mostly used for writing QuickCheck tests.

toIndex

Arguments

:: sh
Shape of the array.

-> sh
Index into the array.

-> Int

Convert an index into its equivalent flat, linear, row-major version.

fromIndex

Arguments

:: sh
Shape of the array.

-> Int
Index into linear representation.

-> sh

Inverse of toIndex.

inShapeRange

Arguments

:: sh
Start index for range.

-> sh
Final index for range.

-> sh
Index to check for.

-> Bool

Check whether an index is within a given shape.

listOfShape :: sh -> [Int]

Convert a shape into its list of dimensions.

shapeOfList :: [Int] -> sh

Convert a list of dimensions to a shape

deepSeq :: sh -> a -> a infixr 0

Ensure that a shape is completely evaluated.

inShape

Arguments

:: Shape sh

=> sh
Shape of the array.

-> sh
Index.

-> Bool

Check whether an index is a part of a given shape.

showShape :: Shape sh => sh -> String

Nicely format a shape as a string