Data.Array.Repa.Slice

Description

Index space transformation between arrays and slices.

Documentation

data All

Select all indices at a certain position.

Constructors

All

data Any sh

Place holder for any possible shape.

Constructors

Any

type family FullShape ss

Map a type of the index in the full shape, to the type of the index in the slice.

type family SliceShape ss

Map the type of an index in the slice, to the type of the index in the full shape.

class Slice ss where

Class of index types that can map to slices.

Minimal complete definition

sliceOfFull, fullOfSlice

Methods

sliceOfFull :: ss -> FullShape ss -> SliceShape ss

Map an index of a full shape onto an index of some slice.

fullOfSlice :: ss -> SliceShape ss -> FullShape ss

Map an index of a slice onto an index of the full shape.